css/selectors/i18n/css4-selectors-lang-006.html

<!DOCTYPE html>
<html  lang="en" >
<head>
<meta charset="utf-8"/>
<title>:lang(en, fr, de), lang="de"</title>
<link rel='author' title='Roger Zanoni' href='mailto:rzanoni@igalia.com'>
<link rel='help' href='https://www.w3.org/TR/selectors-4/#the-lang-pseudo'>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style type='text/css'>
.test div { width: 50px; }
 		#colonlangcontroltest { color: red; font-weight: bold; width: 400px; }
		#colonlangcontroltest:lang(xx) { display:none; }
#box:lang(en, fr, de) { width: 100px; }
</style>
</head>
<body>

<div class="test"><div id="box" lang="de">&#xA0;</div></div>
<p lang='xx' id='colonlangcontroltest'>This test failed because it relies on :lang for results, but :lang is not supported by this browser.

<!--Notes:
This tests a detail related to :lang support. If :lang is not supported, a message will appear and the test will fail.
-->
<script>
test(function() {
assert_equals(document.getElementById('colonlangcontroltest').offsetWidth, 0)
assert_equals(document.getElementById('box').offsetWidth, 100);
}, "If the first language range in :lang argument list doesn't match a lang attribute value, the remaining ranges in the list will be tried.");
</script>

<div id='log'></div>

</body>
</html>
